requirements.txt
streamlit
google-generativeai
geopy
timezonefinder
jyotishyamitra
pytz
pyswisseph
plotly
pandas
pymongo
dnspython
certifi
